+++ engine/include/ecs/component/MeshRenderer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Component.h"
-#include "Assets.h"
+#include "AssetUtilities.h"
 
 namespace nc
 {
