+++ engine/source/assets/AssetHelpers.h
