+++ engine/source/assets/Assets.cpp
@@ -1,4 +1,4 @@
-#include "Assets.h"
+#include "AssetUtilities.h"
 #include "AssetService.h"
 
 namespace nc
