+++ engine/source/graphics/PerFrameRenderState.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Assets.h"
+#include "AssetUtilities.h"
 #include "ecs/component/PointLight.h"
 #include "ecs/Registry.h"
 #ifdef NC_EDITOR_ENABLED
