+++ sample/source/scenes/JareTestScene.cpp
@@ -1,5 +1,5 @@
 #include "JareTestScene.h"
-#include "Assets.h"
+#include "AssetUtilities.h"
 #include "ecs/component/MeshRenderer.h"
 #include "ecs/component/SceneNavigationCamera.h"
 #include "ecs/InvokeFreeComponent.h"
